DRY add component importExportButton
